fix(xai): move native tool markup helper and preserve JSON integers
Address review: keep supporting code under executor/helps, and decode
structured tool arguments with json.Number so large IDs are not rounded.
