+++ internal/runtime/executor/xai_executor_execute.go
@@ -109,7 +109,7 @@ func (e *XAIExecutor) Execute(ctx context.Context, auth *cliproxyauth.Auth, req 
 			if prepared.responseFormat == sdktranslator.FormatOpenAIResponse {
 				out = helps.EnsureResponsesUsageDetails(out)
 			}
-			out = applyXAINativeToolMarkupChatJSON(prepared.responseFormat, out)
+			out = helps.ApplyXAINativeToolMarkupChatJSON(prepared.responseFormat, out)
 			return cliproxyexecutor.Response{Payload: out, Headers: httpResp.Header.Clone()}, nil
 		}
 	}
